requirements.txt
psutil
asciimatics
GPUtil
python-chess==0.31.4
matplotlib
tensorboard
numba
numpy<2.0
requests
pytorch-lightning
pykan
scikit-learn
seaborn
# cupy
torch
scipy